virgl: add fencing interface to virgl_context
Being an interface of virgl_context, this provides per-context fencing.

This is motivated by Vulkan, where each context (i.e., VkInstance) is
independent from one another.  Fences submitted to difference contexts
do not signal in the order they are submitted.  It is in theory and in
practice also true for OpenGL when SW/HW scheduling or context priority
is considered.

Besides being per-context, fences in Vulkan are also per-VkDevice and
per-VkQueue within a context.  This interface uses an opaque uint64_t to
identify the queue a fence is submitted to, similar to how an opaque
uint64_t is used to identify a blob.

Signed-off-by: Chia-I Wu <olvaffe@gmail.com>
Acked-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Louis-Francis Ratté-Boulianne <lfrb@collabora.com>